+++ crates/cargo-gpu/src/lib.rs
@@ -66,14 +66,14 @@ mod linkage;
 mod lockfile;
 mod metadata;
 #[cfg(feature = "naga")]
-mod naga;
+mod naga_transpile;
 mod show;
 mod spirv_source;
 mod test;
 
 pub use install::*;
 #[cfg(feature = "naga")]
-pub use naga::*;
+pub use naga_transpile::*;
 pub use spirv_builder;
 
 /// Central function to write to the user.
